+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Version 0.2.0
 
- - Added support for reading PLY files (https://github.com/w1th0utnam3/splashsurf/pull/1)
+ - Add support for reading PLY files (https://github.com/w1th0utnam3/splashsurf/pull/1)
+ - Update dependencies
+ - Re-export `nalgebra` from lib, use this from binary
 
 ## Version 0.1.0